fix: add support for smaller screen sizes for the Transactions view
