+++ lang/de.json
@@ -682,6 +682,7 @@
   "Crypto": "Crypto",
   "Currency": "Währung",
   "currency": "Währung",
+  "DailyAverage": "Daily avg",
   "Date.BeforeShort": "Before {date, date, short}",
   "Date.DateRange": "{dateFrom, date, short} to {dateTo, date, short}",
   "Date.SinceShort": "Since {date, date, short}",
@@ -2022,6 +2023,12 @@
   "transactions.viewAll": "View All Transactions",
   "TransactionsList.Empty": "No transactions found. <ResetLink>Reset filters</ResetLink> to see all transactions.",
   "TransactionsOverview": "Transactions overview",
+  "TransactionsOverviewSection.CollectiveFilter": "Filter by Collective",
+  "TransactionsOverviewSection.Grants": "Grants",
+  "TransactionsOverviewSection.Invoices": "Invoices",
+  "TransactionsOverviewSection.PeriodFilter": "Filter by Date",
+  "TransactionsOverviewSection.Recurring": "Recurring",
+  "TransactionsOverviewSection.Reimbursements": "Reimbursements",
   "transferwise.error.balance.payment-option-unavailable": "Unable to fund transfer. Please check your balance and try again.",
   "transferwise.error.currencyNotSupported": "This currency is not supported",
   "transferwise.error.default": "An unknown error happened with Wise. Please contact support@opencollective.com.",
