+++ lang/fr.json
@@ -682,6 +682,7 @@
   "Crypto": "Crypto",
   "Currency": "Devise",
   "currency": "devise",
+  "DailyAverage": "Daily avg",
   "Date.BeforeShort": "Before {date, date, short}",
   "Date.DateRange": "{dateFrom, date, short} to {dateTo, date, short}",
   "Date.SinceShort": "Since {date, date, short}",
@@ -2022,6 +2023,12 @@
   "transactions.viewAll": "Voir toutes les transactions",
   "TransactionsList.Empty": "No transactions found. <ResetLink>Reset filters</ResetLink> to see all transactions.",
   "TransactionsOverview": "Transactions overview",
+  "TransactionsOverviewSection.CollectiveFilter": "Filter by Collective",
+  "TransactionsOverviewSection.Grants": "Grants",
+  "TransactionsOverviewSection.Invoices": "Invoices",
+  "TransactionsOverviewSection.PeriodFilter": "Filter by Date",
+  "TransactionsOverviewSection.Recurring": "Recurring",
+  "TransactionsOverviewSection.Reimbursements": "Reimbursements",
   "transferwise.error.balance.payment-option-unavailable": "Unable to fund transfer. Please check your balance and try again.",
   "transferwise.error.currencyNotSupported": "Cette devise n'est pas prise en charge",
   "transferwise.error.default": "Une erreur inconnue s'est produite avec Wise. Veuillez contacter support@opencollective.com.",
