+++ lang/pt-BR.json
@@ -682,6 +682,7 @@
   "Crypto": "Crypto",
   "Currency": "Moeda",
   "currency": "moeda",
+  "DailyAverage": "Daily avg",
   "Date.BeforeShort": "Before {date, date, short}",
   "Date.DateRange": "{dateFrom, date, short} to {dateTo, date, short}",
   "Date.SinceShort": "Since {date, date, short}",
@@ -2022,6 +2023,12 @@
   "transactions.viewAll": "Ver Todas as Transações",
   "TransactionsList.Empty": "Sem transações encontradas. <ResetLink>Redefinir filtros</ResetLink> para ver todas as transações.",
   "TransactionsOverview": "Transactions overview",
+  "TransactionsOverviewSection.CollectiveFilter": "Filter by Collective",
+  "TransactionsOverviewSection.Grants": "Grants",
+  "TransactionsOverviewSection.Invoices": "Invoices",
+  "TransactionsOverviewSection.PeriodFilter": "Filter by Date",
+  "TransactionsOverviewSection.Recurring": "Recurring",
+  "TransactionsOverviewSection.Reimbursements": "Reimbursements",
   "transferwise.error.balance.payment-option-unavailable": "Incapaz de financiar a transferência. Verifique seu saldo e tente novamente.",
   "transferwise.error.currencyNotSupported": "Esta moeda não é suportada",
   "transferwise.error.default": "Um erro desconhecido aconteceu com a Wise. Entre em contato com support@opencollective.com.",
