+++ lang/uk.json
@@ -682,6 +682,7 @@
   "Crypto": "Crypto",
   "Currency": "Валюта",
   "currency": "валюта",
+  "DailyAverage": "Daily avg",
   "Date.BeforeShort": "До {date, date, short}",
   "Date.DateRange": "З {dateFrom, date, short} до {dateTo, date, short}",
   "Date.SinceShort": "З {date, date, short}",
@@ -2022,6 +2023,12 @@
   "transactions.viewAll": "Показати всі транзакції",
   "TransactionsList.Empty": "No transactions found. <ResetLink>Reset filters</ResetLink> to see all transactions.",
   "TransactionsOverview": "Transactions overview",
+  "TransactionsOverviewSection.CollectiveFilter": "Filter by Collective",
+  "TransactionsOverviewSection.Grants": "Grants",
+  "TransactionsOverviewSection.Invoices": "Invoices",
+  "TransactionsOverviewSection.PeriodFilter": "Filter by Date",
+  "TransactionsOverviewSection.Recurring": "Recurring",
+  "TransactionsOverviewSection.Reimbursements": "Reimbursements",
   "transferwise.error.balance.payment-option-unavailable": "Unable to fund transfer. Please check your balance and try again.",
   "transferwise.error.currencyNotSupported": "Ця валюта не підтримується",
   "transferwise.error.default": "An unknown error happened with Wise. Please contact support@opencollective.com.",
