+++ lang/zh.json
@@ -682,6 +682,7 @@
   "Crypto": "加密货币",
   "Currency": "货币",
   "currency": "货币",
+  "DailyAverage": "Daily avg",
   "Date.BeforeShort": "在 {date, date, short} 之前",
   "Date.DateRange": "{dateFrom, date, short} 至 {dateTo, date, short}",
   "Date.SinceShort": "自 {date, date, short}",
@@ -2022,6 +2023,12 @@
   "transactions.viewAll": "View All Transactions",
   "TransactionsList.Empty": "No transactions found. <ResetLink>Reset filters</ResetLink> to see all transactions.",
   "TransactionsOverview": "Transactions overview",
+  "TransactionsOverviewSection.CollectiveFilter": "Filter by Collective",
+  "TransactionsOverviewSection.Grants": "Grants",
+  "TransactionsOverviewSection.Invoices": "Invoices",
+  "TransactionsOverviewSection.PeriodFilter": "Filter by Date",
+  "TransactionsOverviewSection.Recurring": "Recurring",
+  "TransactionsOverviewSection.Reimbursements": "Reimbursements",
   "transferwise.error.balance.payment-option-unavailable": "Unable to fund transfer. Please check your balance and try again.",
   "transferwise.error.currencyNotSupported": "This currency is not supported",
   "transferwise.error.default": "An unknown error happened with Wise. Please contact support@opencollective.com.",
